Updated the UI for tutorial
now will display the keys that are available at the lower left corner of the screen: wasd, and space. After all of them are clicked. They will disappear from the UI. This works as a tutorial to help the user familiarize with the control. However, the client cannot detect input for left and right arrow, so dropped that from the UI.